==> R/99_functions.R <==
# Define project functions ------------------------------------------------


# Define global variables -------------------------------------------------
#vector of order of Location
location_order = c("Upstream", "Wastewater", "Discharge", "Downstream" )
